Project initial commit
The initial commit of this repository adds an afterAll hook which
checks for the current active sync sequence ID. If this ID matches
the latest ID from the registry, it will exit the process cleanly.

Exiting in this fashion is required, so that we can push the AFS
filesystem to the master nodes when it is in known a consistent state.